Created first version of authentication backend.
Note: When an author has been added to the Neo4J database (and thus does
not contain the title attribute), the API breaks down. This is because
we still have to add indexing for ideas and users separately; for now
we still get ALL nodes when listing ideas, which should not be the case.